refactor(ssh): extract SshSession to shared module
Move SshSession struct and SSH_MUX_OPTIONS from backup.rs to
src/ssh_session.rs so it can be reused by the new headscale command.
